pyproject.toml
[project]
name = "finalproject"
version = "0.1.0"
description = ""
authors = [
    {name = "Dexter Corley",email = "dexter.corley@live.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "pillow (>=10.0.0)",
    "pydantic (>=2.5.0)",
    "pdf2image (>=1.17.0)",
    "langfuse (>=3.0.8,<4.0.0)",
    "google-generativeai (>=0.7.0)",
    "transformers (>=4.36.0)",
    "datasets (>=2.16.0)",
    "modal (>=0.62.0)",
    "torch (>=2.1.0)",
    "torchvision (>=0.16.0)",
    "accelerate (>=0.25.0)",
    "python-dotenv (>=1.0.0)",
    "tqdm (>=4.66.0)",
    "numpy (>=1.24.0)",
    "pandas (>=2.0.0)",
    "peft (>=0.15.2,<0.16.0)",
    "matplotlib (>=3.10.3,<4.0.0)",
    "seaborn (>=0.13.2,<0.14.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
